Completed log page and added styling

+++ workout-tracker-backend/src/main/java/com/jok92/workout_tracker_backend/models/auth/SignupDetails.java
@@ -9,6 +9,6 @@ public class SignupDetails implements AuthDetails{
     private String displayName;
     private String email;
     private String password;
-    private Integer age;
+    private Short age;
     private Integer weight;
 }
